// src/shared/generated/LogCaptureRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LogErrorDetails } from "./LogErrorDetails";
import type { LogLevel } from "./LogLevel";
import type { LogSource } from "./LogSource";

export type LogCaptureRecord = { level: LogLevel, source: LogSource, event: string, message: string, contextRawJson?: string, error?: LogErrorDetails, };